guides in notification settings
